% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent.R
\name{adjacent}
\alias{adjacent}
\title{Make Adjacency Object}
\usage{
adjacent(rho)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}
}
\value{
An adjacent \code{propr} object.
}
\description{
This function uses information stored in the \code{@pairs}
 slot to build an adjacency matrix from a \code{propr} object.
}

