% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex.cor.R
\name{aldex.cor}
\alias{aldex.cor}
\title{Correlate CLR with a Continuous Measurement}
\usage{
aldex.cor(clr, conditions)
}
\arguments{
\item{clr}{An \code{aldex.clr} object.}

\item{conditions}{A numeric vector of a continuous variable.}
}
\value{
Returns a data.frame of the average
 Pearson's correlation coefficient (\code{r}),
 Fisher's z-transformation (\code{z}), and p-value
 (\code{p}) for each feature across all
 Monte Carlo instances.
}
\description{
This function uses the Monte Carlo instances from an
 \code{aldex.clr} object to correlate each log-ratio
 transformed feature vector with a continuous
 numeric variable.
}

