% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-aldex2propr.R
\name{aldex.glm}
\alias{aldex.glm}
\title{Build a Generalized Linear Model from CLR Data}
\usage{
aldex.glm(clr, model.matrix, ...)
}
\arguments{
\item{clr}{An \code{aldex.clr} object.}

\item{model.matrix}{A \code{model.matrix}.}

\item{...}{Arguments passed to \code{glm}.}
}
\value{
Returns a data.frame of the average
 coefficients and their p-values for each feature,
 with FDR appended as a \code{BH} column.
}
\description{
This function uses the Monte Carlo instances from an
 \code{aldex.clr} object to build a generalized linear
 model from log-ratio transformed data.
 See \code{\link{lr2glm}}.
}
