% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{alphaTheta}
\alias{alphaTheta}
\title{Calculate alpha Theta}
\usage{
alphaTheta(counts, group, alpha)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the row names in \code{counts}.}

\item{alpha}{A double. See vignette for details. Leave missing
to skip Box-Cox transformation.}
}
\description{
Calculate differential proportionality measure, theta,
 using the Box-Cox transformation method.
}
