% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{alphaThetaW_old}
\alias{alphaThetaW_old}
\title{Calculate Weighted Theta}
\usage{
alphaThetaW_old(counts, group, alpha, weights)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns. Note that this matrix
does not necessarily have to contain counts.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the row names in \code{counts}.}

\item{alpha}{A double. See vignette for details. Leave missing
to skip Box-Cox transformation.}

\item{weights}{A matrix. Pre-computed \code{limma}-based
weights. Optional parameter.}
}
\description{
Do not use this function. For testing purposes only.
}
