% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{as_safe_matrix}
\alias{as_safe_matrix}
\title{Ensure Matrix Has Dim Names}
\usage{
as_safe_matrix(counts)
}
\arguments{
\item{counts}{A data matrix representing counts.}
}
\value{
A matrix with dim names.
}
\description{
Makes sure input data has correct format. For back-end use only.
}
