% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-propd-graflex.R
\name{binTab}
\alias{binTab}
\title{Tabulate Overlap}
\usage{
binTab(A, G)
}
\arguments{
\item{A, G}{A vector or adjacency matrix.}
}
\value{
A table of overlap.
}
\description{
This function tabulates the overlap between
 two vectors or two adjacency matrices.
 It is a faster version of \code{table} that
 only supports binary input.
}
