% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bokeh}
\alias{bokeh}
\title{Make Bokeh Plot}
\usage{
bokeh(rho, k, prompt = TRUE, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{k}{A numeric scalar. The number of groups into which to
cluster the subjects. Clusters calculated  using \code{hclust}.
Optional parameter for \code{\link{bucket}}, \code{\link{prism}},
and \code{\link{bokeh}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the courtesy prompt when working with big data.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}
}
\value{
Returns cluster membership if \code{k} is provided.
 Otherwise, returns a \code{ggplot} object.
}
\description{
Plots the feature variances for each log-ratio transformed
 feature pair in the \code{propr} object. Highly proportional
 pairs will aggregate near the \code{y = x} diagonal.
 Clusters that appear toward the top-right of the
 figure contain features with highly variable abundance across
 all samples. Clusters that appear toward the
 bottom-left of the figure contain features with fixed
 abundance across all samples. Uses a log scale.
}
\details{
Providing the argument \code{k} will color feature pairs
 by co-cluster membership. In other words, a feature pair
 will receive a color if and only if both features belong
 to same the cluster (calculated using \code{hclust}).
}

