% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{calculateFDR}
\alias{calculateFDR}
\title{Calculate FDR Cutoffs}
\usage{
calculateFDR(theta, ptheta, cutoff = seq(0.6, 0.9, 0.1))
}
\arguments{
\item{theta}{A \code{data.frame}. The result of a
\code{calculateTheta} call.}

\item{ptheta}{A \code{data.frame}. The result of a
\code{permuteTheta} call.}

\item{cutoff}{For \code{updateCutoffs}, a numeric vector.
this argument provides the FDR cutoffs to test.
For graph functions, a numeric scalar. This argument
indicates the maximum theta to include in the figure.
For graph functions, a large integer will instead
retrieve the top N pairs as ranked by theta.}
}
\description{
This function uses the result of a \code{permuteTheta} call
 to calculate false discovery rate (FDR) cutoffs.
}
