% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-propd-graflex.R
\name{calculateOR}
\alias{calculateOR}
\title{Calculate Odds Ratio}
\usage{
calculateOR(A, G)
}
\description{
This function calculates an odds ratio for the
 edge overlap between two non-random graphs.
}
\details{
Note that this function calculates overlap for the
 lower-left triangle of the input matrices.
}
