% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated2.R
\name{calculateTheta_old}
\alias{calculateTheta_old}
\title{Calculate Theta}
\usage{
calculateTheta_old(counts, group)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the row names in \code{counts}.}
}
\description{
Do not use this function. For testing purposes only.
}
