% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{calculateThetaW_old}
\alias{calculateThetaW_old}
\title{Calculate Weighted Theta}
\usage{
calculateThetaW_old(counts, group)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns. Note that this matrix
does not necessarily have to contain counts.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the row names in \code{counts}.}
}
\description{
Do not use this function. For testing purposes only.
}
