% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\docType{data}
\name{caneToad.counts}
\alias{caneToad.counts}
\title{Example Cane Toad Count Data}
\format{An object of class \code{matrix} with 20 rows and 25774 columns.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(caneToad.counts)
}
\description{
Raw RNA-seq counts for twenty toads sampled from one
 of two regions in Australia. Data set reduced to
 exclude any transcripts without at least 10 counts
 in at least 10 samples.
}
\keyword{datasets}
