% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{caneToad.groups}
\alias{caneToad.groups}
\title{Example Cane Toad Group Data}
\format{An object of class \code{character} of length 20.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(caneToad.groups)
}
\description{
Example Cane Toad Group Data
}
\keyword{datasets}

