% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-functions.R
\name{clo}
\alias{clo}
\title{Closure}
\usage{
clo(X, check = FALSE)
}
\arguments{
\item{X}{A matrix or dataframe of positive numeric values}

\item{check}{A logical scalar}
}
\value{
A version of \code{X} where each row has been scaled so they sum to 1.
}
\description{
\code{clo} divides each row of \code{X} by its row sum
}
\details{
If \code{check} is \code{TRUE} then this function will stop if
there are any negative or \code{NA} values in \code{X}
}
\examples{
X <- matrix(1:12, nrow=3)
x <- clo(X)
rowSums(x)
}
