% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-functions.R
\name{clr}
\alias{clr}
\title{Centred logratio transformation}
\usage{
clr(X, check = FALSE)
}
\arguments{
\item{X}{A matrix or dataframe of positive numeric values}

\item{check}{A logical scalar}
}
\value{
The logarithm of \code{X} where each row has been shifted to have mean 0.
\deqn{\mathrm{clr}(x) = \log x_i - \sum_{i=1}^D \log x_i}
}
\description{
\code{clr} takes the log of each row of X and centres it (i.e., subtracts the mean).
}
\details{
If \code{check} is \code{TRUE} then this function will stop if
there are any negative or \code{NA} values in \code{X}
}
\examples{
X <- matrix(1:12, nrow=3)
x <- clr(X)
rowSums(x)             # Pretty close to zero
apply(exp(x), 1, prod) # The row products of exp(x) will be 1
}
