% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-propr-backend.R
\name{compute_iqlr}
\alias{compute_iqlr}
\title{Compute IQLR (Interquartile Range Log Ratio) Features}
\usage{
compute_iqlr(counts)
}
\arguments{
\item{counts}{A data matrix representing counts.
It is assumed that the matrix contains numerical values only.}
}
\value{
A numeric vector representing the indices of features selected by IQLR.
}
\description{
This function computes the IQLR features from the input count matrix.
 The IQLR is based on the log-ratio transformation of the counts, and
 it selects features with variance values falling within the
 interquartile range.
}
\examples{
# Sample input count data
data <- matrix(c(1, 2, 3, 4, 0, 6), nrow = 2, byrow = TRUE)

# Compute IQLR features
iqlr_features <- compute_iqlr(data)

}
