% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-updateCutoffs.R
\name{countValuesBeyondThreshold}
\alias{countValuesBeyondThreshold}
\title{Count Values Greater or Less Than a Threshold}
\usage{
countValuesBeyondThreshold(values, cutoff, direct)
}
\arguments{
\item{values}{A numeric vector.}

\item{cutoff}{A numeric value.}

\item{direct}{A logical value. If \code{TRUE}, direct relationship is considered.}
}
\value{
The number of values greater or less than the threshold.
}
\description{
This function counts the number of values greater or less than a threshold.
The direction depends on if a direct or inverse relationship is asked,
as well as the sign of the threshold.
}
