% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{cytescape}
\alias{cytescape}
\title{Export Indexed Pairs}
\usage{
cytescape(object, minPairs = 2)
}
\arguments{
\item{object}{An object of class \code{propr}.}

\item{minPairs}{An integer scalar. Subsets the interaction
network to exclude any pair without a node that participates
in at least this many total pairs.}
}
\value{
Returns a \code{data.frame} of indexed pairs.
}
\description{
This function exports a \code{data.frame} of indexed pairs
 along with their corresponding values from \code{object@matrix}.
 In doing so, this function provides a preview of the
 interaction network, built using \code{igraph}.
 We recommend using this result as input to a standalone
 network visualization tool like Cytoscape.
}

