% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{dendroCheck}
\alias{dendroCheck}
\title{Dendrogram Plot Check}
\usage{
dendroCheck()
}
\description{
Performs data checks before plotting, triggering messages
 or errors when appropriate. For back-end use only.
}
