% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{differentialCheck}
\alias{differentialCheck}
\title{Differential Proportionality Check}
\usage{
differentialCheck(x, y, forceBoth)
}
\arguments{
\item{x}{A \code{propr} object.}

\item{y}{A \code{propr} object. Optional.}

\item{forceBoth}{Toggles whether to perform checks for the
second \code{propr} object.}
}
\description{
Performs data checks when comparing two \code{propr} objects,
 triggering messages or errors when appropriate. For back-end
 use only.
}

