% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-propd-genewise.R
\name{get_connectivity}
\alias{get_connectivity}
\title{Get connectivity for each gene based on pairwise propd results}
\usage{
get_connectivity(propd, pairwise_fdr = 0.05)
}
\arguments{
\item{propd}{A \code{\link{propd}} object, with FDR values from updateF. Note:
for the moment, only theta results with F-stats are supported. Later we will look 
on the option to get FDR values based on permutations.}

\item{pairwise_fdr}{}
}
\value{
A data frame with genewise connectivity results.
}
\description{
It computes the per-gene connectivity by counting the number of
significant pairwise relationships each gene has. It also provides
a genewise significance stat (for the moment we provide a fake one
by averaging all the pairwise FDR values for each gene, but later
we will look into more robust methods to get genewise p-values).
}
