% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2a-propd-backend.R
\name{get_theta_matrix}
\alias{get_theta_matrix}
\title{Parse propd theta values to a matrix}
\usage{
get_theta_matrix(propd)
}
\value{
A symmetric matrix.
}
\description{
This function returns a symmetric matrix of \code{propd} values.
Note that the active theta values are used for the matrix. The 
diagonal is set to NA.
}
