% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-propd-genewise.R
\name{get_weighted_connectivity}
\alias{get_weighted_connectivity}
\title{Get weighted connectivity for each gene based on pairwise propd results}
\usage{
get_weighted_connectivity(propd, pairwise_fdr = 0.05)
}
\arguments{
\item{propd}{A \code{\link{propd}} object, with FDR values from updateF. Note:
for the moment, only theta results with F-stats are supported. Later we will look 
on the option to get FDR values based on permutations.}

\item{pairwise_fdr}{}
}
\value{
A data frame with genewise weighted connectivity results.
}
\description{
It computes the per-gene weighted connectivity by summing the strength 
of significant pairwise relationships each gene has. The strength of a 
relationship can be defined #' as the inverse of the theta value (i.e., 1/theta) 
(i.e., 1/theta) for significant pairs, which gives more weight to stronger 
relationships. It also provides a genewise significance stat by averaging 
all the pairwise FDR values for each gene.
}
