% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getAdj.R
\name{getAdj}
\alias{getAdj}
\title{Get Object as Adjacency Matrix}
\usage{
getAdj(object, cutoff = 1, above = FALSE)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}
}
\value{
An adjacency matrix.
}
\description{
This function uses \code{getMatrix} to return a \code{propr}
or \code{propd} object as an adjacency matrix. The diagonal
is set to 0.
}
