% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getAdjacency.R
\name{getAdjacencyFstat}
\alias{getAdjacencyFstat}
\title{Get Adjacency Matrix as indicated by F-statistics}
\usage{
getAdjacencyFstat(object, pval = 0.05, fdr_adjusted = TRUE)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{pval}{A float value for the p-value. Default is 0.05.}

\item{fdr_adjusted}{A boolean. If TRUE, use the the FDR- adjusted p-values. 
Otherwise, get significant pairs based on the theoretical F-statistic cutoff.}
}
\value{
An adjacency matrix.
}
\description{
This function gets the significant pairs, according to the F-statistics. 
Then it fills the adjacency matrix with 1 if pair is significant, otherwise 0.
Note that it can only be applied to theta_d, as updateF only works for theta_d.
}
