% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getAdj.R
\name{getAdjFstat}
\alias{getAdjFstat}
\title{Get Adjacency Matrix as indicated by F-statistics}
\usage{
getAdjFstat(object, pval = 0.05, fdr = TRUE)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{pval}{A float value for the p-value. Default is 0.05.}

\item{fdr}{A boolean. If TRUE, use the the FDR- adjusted p-values. Otherwise,
get significant pairs based on the theoretical F-statistic cutoff.}
}
\value{
An adjacency matrix.
}
\description{
This function gets the significant pairs, according to the F-statistics. 
Then it fills the adjacency matrix with 1 if pair is significant, otherwise 0.
The diagonal is set to 0.
}
