% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-propd-graflex.R
\name{getFDR}
\alias{getFDR}
\title{Calculate Odds Ratio FDR}
\usage{
getFDR(actual, permuted)
}
\arguments{
\item{actual}{A result from \code{\link{calculateOR}}.}

\item{permuted}{A result from \code{\link{permuteOR}}.}
}
\value{
A \code{data.frame} of the FDRs for over-
 and under- enrichment.
}
\description{
This function calculates the false discovery rate (FDR)
 for over- and under-enrichment by counting the number of
 times the actual OR was greater than
 (or less than) a permuted OR.
}
