% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-updateCutoffs.R
\name{getFdrRandcounts.propd.run}
\alias{getFdrRandcounts.propd.run}
\title{This function counts the permuted values greater or less than each cutoff,
using a single core, for a propd object.}
\usage{
getFdrRandcounts.propd.run(object, cutoffs)
}
\description{
This function counts the permuted values greater or less than each cutoff,
using a single core, for a propd object.
}
