% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-updateCutoffs.R
\name{getFdrRandcounts.propr.parallel}
\alias{getFdrRandcounts.propr.parallel}
\title{This function counts the permuted values greater or less than each cutoff,
using parallel processing, for a propr object}
\usage{
getFdrRandcounts.propr.parallel(object, cutoffs, ncores)
}
\description{
This function counts the permuted values greater or less than each cutoff,
using parallel processing, for a propr object
}
