% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getMatrix.R
\name{getMatrix}
\alias{getMatrix}
\title{Get Matrix from Object}
\usage{
getMatrix(object)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}
}
\value{
A matrix.
}
\description{
This function provides a unified wrapper to retrieve a matrix
 of \code{propr} or \code{propd} values.
}
