% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getCutoff.R
\name{getMovingAverage}
\alias{getMovingAverage}
\title{Caclulate the moving average of a vector.}
\usage{
getMovingAverage(values, window_size = 1)
}
\arguments{
\item{values}{A numeric vector.}

\item{window_size}{An integer. The size of the window to calculate the
moving average. Default is 1.}
}
\description{
Caclulate the moving average of a vector.
}
