% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-getResults.R
\name{getNetwork}
\alias{getNetwork}
\title{Get Network from Object}
\usage{
getNetwork(propr.object, propr.cutoff = NA, thetad.object,
  thetad.cutoff = NA, thetae.object, thetae.cutoff = NA, col1, col2,
  d3 = FALSE)
}
\arguments{
\item{propr.object, thetad.object, thetae.object}{A \code{propr}
object or an appropriate \code{propd} object.}

\item{propr.cutoff, thetad.cutoff, thetae.cutoff}{A cutoff
argument passed to \code{\link{getResults}}.}

\item{col1}{A character vector. Specifies which nodes
to color \code{red} or \code{blue}, respectively.}

\item{col2}{A character vector. Specifies which nodes
to color \code{red} or \code{blue}, respectively.}

\item{d3}{A boolean. Use \code{rgl} to plot 3D network.}
}
\value{
A network object.
}
\description{
This function provides a unified wrapper to build networks
 from \code{propr} and \code{propd} objects.
}
