% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-propd-graflex.R
\name{getOR}
\alias{getOR}
\title{Calculate Odds Ratio}
\usage{
getOR(A, G)
}
\value{
A \code{data.frame} of results.
}
\description{
This function calculates the overlap between
 two vectors or two adjacency matrices.
 It returns the OR as well as other metrics.
}
