% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-updateCutoffs.R
\name{getPermutedTheta}
\alias{getPermutedTheta}
\title{Get the theta values for a given permutation}
\usage{
getPermutedTheta(object, k)
}
\description{
Get the theta values for a given permutation
}
