% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-updateCutoffs.R
\name{getPermutedThetaMod}
\alias{getPermutedThetaMod}
\title{Get the theta mod values for a given permutation}
\usage{
getPermutedThetaMod(object, k)
}
\description{
Get the theta mod values for a given permutation
}
