% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getRatios.R
\name{getRatios}
\alias{getRatios}
\title{Get (Log-)ratios from Object}
\usage{
getRatios(object, switch = TRUE)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}

\item{switch}{A boolean. For \code{propd}, toggles whether all ratios
should have same orientation with respect to log-ratio means.}
}
\value{
A \code{data.frame} of (log-)ratios.
}
\description{
This function provides a unified wrapper to retrieve (log-)ratios
 from \code{propr} and \code{propd} objects.
}
\details{
When the \code{propd} object is made using \code{alpha}, this function returns
 the (log-)ratios as \code{(partner^alpha - pair^alpha) / alpha}.
}
