% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-get.R
\name{getReference}
\alias{getReference}
\title{Get Reference to Approximate CLR}
\usage{
getReference(counts, alpha = NA)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns. Note that this matrix
does not necessarily have to contain counts.}

\item{alpha}{A double. See vignette for details. Leave missing
to skip Box-Cox transformation.}
}
\value{
A reference as a character or integer.
}
\description{
This function finds the reference that is most
 proportional to the geometric mean of the samples.
 Using this reference for an alr-transformation
 will permit an analysis that resembles that of a
 clr-transformation, but arguably with greater
 interpretability.
}
