% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getResults.R
\name{getSignificantResultsFDR}
\alias{getSignificantResultsFDR}
\alias{getSignificantResultsFDR.propd}
\alias{getSignificantResultsFDR.propr}
\title{Get Significant Results from Object based on the permutation tests.}
\usage{
getSignificantResultsFDR(object, fdr = 0.05, window_size = 1, tails = NULL)

getSignificantResultsFDR.propd(object, fdr = 0.05, window_size = 1)

getSignificantResultsFDR.propr(
  object,
  fdr = 0.05,
  window_size = 1,
  tails = NULL
)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{fdr}{A float value for the false discovery rate. Default is 0.05.}

\item{window_size}{An integer. Default is 1. When it is greater than 1, the FDR
values would be smoothed out by a moving average of the given window size.}

\item{tails}{1 for one-sided on the right, and 2 for two-sided. When NULL, use default
test for the given metric. This is only relevant for \code{propr} objects.}
}
\value{
A \code{data.frame} of results.
}
\description{
This function provides a unified wrapper to retrieve results
 from a \code{propr} or \code{propd} object keeping only the
 statistically significant pairs.
}
\section{Methods}{

\code{getSignificantResultsFDR.propd:}
This function retrieves results from a \code{propd} object keeping 
only the statistically significant pairs.


\code{getSignificantResultsFDR.propr:}
This function retrieves results from a \code{propr} object keeping 
only the statistically significant pairs.
}

