% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getResults.R
\name{getSignificantResultsFstat}
\alias{getSignificantResultsFstat}
\title{Get Significant Results based on the F-stats.}
\usage{
getSignificantResultsFstat(object, pval = 0.05, fdr = TRUE)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}

\item{pval}{A numeric. The p-value to use.}

\item{fdr}{A boolean. If TRUE, the function will return the
significant pairs based on the FDR adjusted p-values. Otherwise,
it will return the significant pairs based on the F-statistic cutoff.}
}
\value{
A \code{data.frame} of results.
}
\description{
This function provides a unified wrapper to retrieve results
 from a \code{propd} object keeping only the statistically 
 significant pairs. Note that it can only be applied to theta_d.
}
