% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{ggdend}
\alias{ggdend}
\title{Dendrogram Plot Wrapper}
\usage{
ggdend(dendrogram)
}
\arguments{
\item{dendrogram}{A result from \code{as.dendrogram}.}
}
\description{
Builds \code{ggplot2} dendrograms. For back-end use only.
}

