% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-proprCall.R
\name{ivar2index}
\alias{ivar2index}
\title{Build Index from ivar Argument}
\usage{
ivar2index(counts, ivar)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns. Note that this matrix
does not necessarily have to contain counts.}

\item{ivar}{A numeric scalar. Specifies reference feature(s)
for additive log-ratio transformation. The argument will also
accept feature name(s) instead of the index position(s).
Set to "iqlr" to use inter-quartile log-ratio transformation.
Ignore to use centered log-ratio transformation.}
}
\value{
A numeric vector of indices for the reference set.
}
\description{
This function builds an index from the \code{ivar} argument. Used by
 the \code{propr} initialize method and \code{updateF}.
}
