% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-propr-backend.R
\name{logratio}
\alias{logratio}
\title{Perform log-ratio transformation}
\usage{
logratio(counts, ivar, alpha = NA)
}
\arguments{
\item{counts}{A data matrix representing counts.
It is assumed that the matrix contains numerical values only.}

\item{ivar}{An indicator specifying the method for log-ratio transformation.
It can take the following values:
 - "clr" (default): Centered log-ratio transformation.
 - "alr": Additive log-ratio transformation ("pcor.bshrink" metric only).
 - "iqlr": Inter-quartile log-ratio transformation from ALDEx2.
 - The explicit name(s) or index(es) of variable(s) to use as a reference.
 - Use NA to skip log-ratio transformation and any other pre-processing, like
zero replacement. This is useful when the input data is already pre-processed.}

\item{alpha}{The alpha parameter used in the alpha log-ratio transformation.}
}
\value{
A matrix with log-ratio transformed values.
}
\description{
This function applies a log-ratio transformation to a given data matrix
 with or without using an alpha parameter.
}
\examples{
# Sample counts matrix
counts_matrix <- matrix(c(10, 20, 30, 40, 50, 60, 70, 80, 90), nrow = 3, byrow = TRUE)
colnames(counts_matrix) <- c("A", "B", "C")
rownames(counts_matrix) <- c("Sample1", "Sample2", "Sample3")

# Perform log-ratio transformation without alpha
logratio(counts_matrix, ivar = "A")

# Perform log-ratio transformation with alpha 0.5
logratio(counts_matrix, ivar = "A", alpha = 0.5)

# Skip log-ratio transformation
logratio(counts_matrix, ivar = NA)

}
