% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-propr-backend.R
\name{logratio_without_alpha}
\alias{logratio_without_alpha}
\title{Perform log-ratio transformation without alpha parameter}
\usage{
logratio_without_alpha(ct, use)
}
\arguments{
\item{ct}{A data matrix for which the log-ratio transformation will be
performed. It is assumed that the matrix contains numerical values only.}

\item{use}{An integer vector specifying the subset of columns to be used
for the log-ratio transformation.}
}
\value{
A matrix containing the log-ratio transformed data.
}
\description{
This function applies a log-ratio transformation to a given data matrix
 without using an alpha parameter. The log-ratio transformation is based
 on a selected subset of columns specified by the `use` argument.
}
\examples{
# Sample input data
data <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2, byrow = TRUE)

# Applying log-ratio transformation to rows using columns 2 and 3 as reference
result <- logratio_without_alpha(data, use = c(2, 3))

}
