% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-aldex2propr.R
\name{lr2glm}
\alias{lr2glm}
\title{Build a Generalized Linear Model from CLR Data}
\usage{
lr2glm(lr, model.matrix, ...)
}
\arguments{
\item{lr}{A data.frame. A log-ratio transformed counts matrix.}

\item{model.matrix}{A \code{model.matrix}.}

\item{...}{Arguments passed to \code{glm}.}
}
\value{
Returns a data.frame of the estimated
 coefficients and their p-values for each feature,
 with FDR appended as a \code{BH} column.
}
\description{
This function builds a generalized linear model from
 log-ratio transformed data.
}
