% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\docType{data}
\name{mail}
\alias{mail}
\title{Mock Mail Count Data}
\format{An object of class \code{matrix} with 5 rows and 4 columns.}
\usage{
data(mail)
}
\description{
Includes mock count data for 5 days and 4 zip codes.
}
\keyword{datasets}
