% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\docType{data}
\name{marg.abs}
\alias{marg.abs}
\title{Example Absolute mRNA}
\format{An object of class \code{data.frame} with 16 rows and 3031 columns.}
\usage{
data(marg.abs)
}
\description{
Data generated with supplemental script provided by
 <DOI:10.1371/journal.pcbi.1004075>. Data originally
 sourced from <DOI:10.1016/j.cell.2012.09.019>.
 A time series of yeast mRNA abundance after removal
 of a key nutrient. Absolute abundance estimated
 by multiplying microarray signal (relative to first
 time point) by the initial nCounter-calibrated and
 copy-per-cell-adjusted RNA-seq abundance (averaged
 across two replicates). Divide absolute abundances
 by total sample abundance to make data relative.
}
\keyword{datasets}
