% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mds}
\alias{mds}
\title{Make MDS Plot}
\usage{
mds(rho, group)
}
\description{
Plots the first two principal components as calculated
 using the log-ratio transformed feature vectors. This
 provides a statistically valid alternative to
 conventional principal components analysis (PCA) and
 multi-dimensional scaling (MDS) plotting.
}
\details{
For more information, see DOI:10.1139/cjm-2015-0821.
}

