% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraph.R
\name{migraph}
\alias{migraph}
\alias{migraph.add}
\alias{migraph.clean}
\alias{migraph.color}
\title{\code{igraph} Helper Functions}
\usage{
migraph.add(g, names1, names2, force = TRUE)

migraph.color(g, names1, names2, col)

migraph.clean(g)
}
\arguments{
\item{g}{An \code{igraph} object.}

\item{names1, names2}{A character vector. The \code{names1}
argument defines a first set of vertices. The \code{names2}
argument defines a second set of vertices to which the
first set connects (i.e., element-wise), thereby defining
a set of edges.}

\item{force}{A boolean. If true, the function adds any
missing vertcies before adding edges. If false, the
function only adds edges that have both vertices
already present.}

\item{col}{A character string. The color applied to all
vertices (or edges) specified by the \code{names1} (or
\code{names2}) argument.}
}
\value{
An \code{igraph} object.
}
\description{
\code{igraph} Helper Functions
}

