% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrmodel.R
\name{modelCLR}
\alias{modelCLR}
\title{Encapsulate a clr-Transformation}
\usage{
modelCLR(x, MARGIN = 1)
}
\arguments{
\item{x}{A non-negative count matrix.}

\item{MARGIN}{If \code{MARGIN = 1}, rows contain sample data.
If \code{MARGIN = 2}, columns contain sample data.}
}
\description{
This function encapsulates the rule for a clr-transformation to apply
 to an independent dataset (via \code{predict}).
}
\seealso{
\code{\link{lrmodel-class}}
}

