% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{multiplot}
\alias{multiplot}
\title{Plot Multiple Graphs}
\usage{
multiplot(..., cols = 1)
}
\arguments{
\item{...}{Multiple plots.}

\item{cols}{A numeric scalar. The number of plot columns.}
}
\description{
Easily plot multiple graphs within the same window. Code adapted from
 http://www.cookbook-r.com/. For back-end use only.
}
