% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{packageCheck}
\alias{packageCheck}
\title{Package Check}
\usage{
packageCheck(package)
}
\arguments{
\item{package}{A character string. An R package.}
}
\description{
Checks whether the user has the required package installed.
 For back-end use only.
}

