% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-propr-backend.R
\name{pcor.bshrink}
\alias{pcor.bshrink}
\title{Basis Covariance Shrinkage and Partial Correlation Calculation}
\usage{
pcor.bshrink(ct, outtype = c("clr", "alr"))
}
\arguments{
\item{ct}{A data matrix representing the counts. Each row should represent
observations, and each column should represent different variables.}

\item{outtype}{A character vector specifying the output type. It can take
either "clr" (centered log-ratio) or "alr" (additive log-ratio). Since
the reference does not affect the logratio partial correlation coefficients,
the index of the reference is not needed for "alr". Also, "clr" is recommended
because of the same reason, at the same time avoiding losing one dimension.}
}
\value{
A matrix representing the shrunk partial correlation matrix.
}
\description{
This function performs covariance shrinkage on the basis matrix
 and calculates the partial correlation matrix. The function can
 output the results in two formats: centered log-ratio (clr) or
 additive log-ratio (alr).
}
\examples{
# Sample input count data
data <- iris[,1:4]

# Calculate partial correlation matrix using clr transformation
result_clr <- pcor.bshrink(data, outtype = "clr")

# Calculate partial correlation matrix using alr transformation
result_alr <- pcor.bshrink(data, outtype = "alr")

}
