% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\docType{data}
\name{pd.e}
\alias{pd.e}
\title{Example propd Object}
\format{An object of class \code{propd} of length 1.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(pd.e)
}
\description{
Includes results from \code{\link{propd}} as applied to
 cane toad transcripts with at least 40 counts in at least
 20 samples (after removing any transcripts with 0 counts).
 The resultant object is filtered to include only the top
 1000 theta_e values in the \code{@results} slot.
}
\keyword{datasets}
