% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-perb.R
\name{perb}
\alias{perb}
\title{Calculate proportionality metric p.}
\usage{
perb(counts, ivar = 0, iter = 0, iterSize = nrow(counts) - (ivar > 0),
  iterHow = 1, onlyDistr = FALSE)
}
\arguments{
\item{counts}{A data.frame or matrix. See "count matrix" described in \code{\link{phit}} function details.}

\item{ivar}{A numeric scalar. Specificies feature to use as reference for additive log-ratio transformation.}

\item{iter}{A numeric scalar. Fits \code{iter*iterSize*(iterSize-1)/2} values to an empiric distribution. Skip with \code{iter = 0}.}

\item{iterSize}{A numeric scalar. Fits \code{iter*iterSize*(iterSize-1)/2} values to an empiric distribution.}

\item{iterHow}{A numeric scalar. Select \code{1} to randomize feature vectors or \code{2} to randomize subject vectors.}

\item{onlyDistr}{A logical. Provided for backend use. Evokes function to return only \code{ecdf} fit.}
}
\value{
Returns a \code{propr} object.
}
\description{
\code{perb} returns a \code{propr} object containing measures of proportionality.
}
\details{
Calculates proportionality metric p described in Lovell 2015 and expounded in Erb 2016. Uses
centered log-ratio transformation of data by default, but will use additive log-ratio transformation
of data if non-zero \code{ivar} provided.
}
\examples{
randomNum <- sample(1:1000, size = 2000 * 22, replace = TRUE)
counts <- matrix(randomNum, nrow = 2000, ncol = 22)
prop <- perb(counts, ivar = 0, iter = 0)
}
\seealso{
\code{\link{propr-class}}, \code{\link{phit}}
}

