% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-propd-graflex.R
\name{permuteOR}
\alias{permuteOR}
\title{Permute Odds Ratio}
\usage{
permuteOR(A, G, p = 500)
}
\arguments{
\item{A, G}{An adjacency matrix.}

\item{p}{An integer. The number of overlaps to permute.}
}
\description{
This function permutes \code{p} odds ratios for the
 edge overlap between two non-random graphs.
 It does this by randomly shuffling the rows and
 columns of \code{A} (jointly, thus preserving
 the degree distribution).
}
\details{
Note that this function calculates overlap for the
 lower-left triangle of the input matrices.
}
