% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{permuteTheta_old}
\alias{permuteTheta_old}
\title{Permute Theta}
\usage{
permuteTheta_old(counts, group, p)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns. Note that this matrix
does not necessarily have to contain counts.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the row names in \code{counts}.}

\item{p}{An integer. The number of permutation cycles.}
}
\description{
Do not use this function. For testing purposes only.
}
