% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-functions.R
\name{phiDF}
\alias{phiDF}
\title{Pairwise proportionality, slope and other statistics of columns}
\usage{
phiDF(X)
}
\arguments{
\item{X}{A matrix or dataframe}
}
\value{
A list of three elements \code{b}, \code{p} and \code{r2}.
}
\description{
\code{phiDF} returns a dataframe of various statistics for each pair of columns in
\code{X}.
}
\details{
\strong{Note:} \code{X} should be the result of a centred logratio transformation
}
\examples{
N <- 100 # Number of observations
# Make a data frame with columns a and b roughly proportional
# and columns c and d unrelated
a <- seq(from=5, to=15, len=N)
b <- a * rnorm(N, 1, 0.1)
c <- rnorm(N,10)
d <- rnorm(N,10)
X <- data.frame(a, b, c, d)
pairs(X)
pairs(clr(X)) # Note the spurious correlation between variables c and d
phiDF(X)
# Note that phi and phisym are related to the slope and r^2 values:
with(phiDF(X),
  all.equal(
    phisym,
    (1 + b^2 - 2*b*sqrt(r2))/(1 + b^2 + 2*b*sqrt(r2))
  )
)
with(phiDF(X),
  all.equal(
    phi,
    (1 + b^2 - 2*b*sqrt(r2))
  )
)
}
