% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-functions.R
\name{phisym}
\alias{phisym}
\title{Symmetric phi statistic}
\usage{
phisym(X)
}
\arguments{
\item{X}{A matrix or dataframe}
}
\value{
TBA.
}
\description{
\code{phisym} returns a matrix where element (i,j) is
the symmetric phi statistic between columns i and j of \code{X}.
}
\details{
\code{X} should be the result of a centred logratio transformation
}
\examples{
N <- 10 # Number of observations
# Make a data frame with columns a and b roughly proportional
# and columns c and d roughly proportional
X <- data.frame(a=(1:N), b=(1:N) * rnorm(N, 10, 0.1),
                c=(N:1), d=(N:1) * rnorm(N, 10, 1.0))
round(phisym(clr(X)),2)
}
