% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perb.R
\name{phit}
\alias{phit}
\title{Calculate phi}
\usage{
phit(counts, symmetrize = TRUE)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with subjects as rows and features as columns.}

\item{symmetrize}{A logical. If \code{TRUE}, forces symmetry by reflecting the "lower left triangle".}
}
\value{
Returns a \code{propr} object.
}
\description{
\code{phit} returns a \code{propr} object containing a measure of proportionality, phi.
}
\details{
Let D represent any number of features measured across N biological replicates
	subjected to a binary or continuous event E. For example, E could represent case-control
	status, treatment status, treatment dose, or time. This function converts a
	"count matrix" with N rows and D columns into a proportionality matrix of D rows and D
	columns containing a value of phi for each feature pair. One can think of the resultant
	matrix as analogous to a distance matrix, except that it has no symmetry unless forced.
}
\examples{
library(propr)
data(mail)
phi <- phit(mail, symmetrize = TRUE)
}
\seealso{
\code{\link{propr}}, \code{\link{propr-class}}, \code{\link{perb}}
}

