% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{plotCheck}
\alias{plotCheck}
\title{Plot Check}
\usage{
plotCheck(rho, prompt, plotly, indexNaive)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the prompt when plotting a large number of features.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package. This will also
have the function return the internal \code{ggplot} object(s),
useful for customization of the figure.}

\item{indexNaive}{Toggles whether to perform checks for an
"index-naive" plot function.}
}
\description{
Performs data checks before plotting, triggering messages
 or errors when appropriate. For back-end use only.
}

