% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-posthoc.R
\name{posthoc}
\alias{posthoc}
\title{Perform Post-Hoc Testing}
\usage{
posthoc(object)
}
\arguments{
\item{object}{A \code{\link{propd}} object.}
}
\value{
A \code{\link{propd}} object.
}
\description{
After running an ANOVA of more than 2 groups, it is useful
 to know which of the groups differ from the others. This
 question is often answered with post-hoc testing. This
 function implements post-hoc pairwise differential
 proportionality analyses for more than 2 groups.
}
\details{
The ANOVA p-values are adjusted once (column-wise) during
 the original multi-group analysis. The post-hoc p-values
 are adjusted once (row-wise) for the number
 of post-hoc tests. The post-hoc adjustment
 is p times the number of post-hoc tests.

Please note that a significant post-hoc test without
 a significant ANOVA test is never significant!
}
