% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{progress}
\alias{progress}
\title{Make Progress Bar}
\usage{
progress(i, k, numTicks)
}
\arguments{
\item{i}{The current iteration.}

\item{k}{Total iterations.}

\item{numTicks}{The result of \code{progress}.}
}
\value{
The next \code{numTicks} argument.
}
\description{
Make Progress Bar
}
