% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{promptCheck}
\alias{promptCheck}
\title{Feature Check}
\usage{
promptCheck(N)
}
\arguments{
\item{N}{An integer. The number of features.}
}
\description{
Prompts user when performing an operation on an unusually
 large set of features. For back-end use only.
}
