% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-propd-OOP.R, R/2-propd-help.R, R/2-propd.R,
%   R/2b-propd-frontend.R
\docType{class}
\name{propd-class}
\alias{propd-class}
\alias{show,propd-method}
\alias{propd}
\alias{setDisjointed}
\alias{setEmergent}
\alias{setActive}
\alias{updateF}
\title{The propd Method}
\usage{
\S4method{show}{propd}(object)

propd(counts, group, alpha = NA, p = 0, fixseed = FALSE, weighted = FALSE)

setDisjointed(propd)

setEmergent(propd)

setActive(propd, what = "theta_d")

updateF(propd, moderated = FALSE, ivar = "clr")
}
\arguments{
\item{object}{A \code{propd} object.}

\item{counts}{A data matrix representing counts.
It is assumed that the matrix contains numerical values only.}

\item{group}{A character vector representing group labels indicating the
assignment of each count to different groups.}

\item{alpha}{The alpha parameter used in the alpha log-ratio transformation.}

\item{p}{The number of permutations to perform for calculating the false
discovery rate (FDR). The default is 0.}

\item{fixseed}{A logical value indicating whether to fix the random seed 
for generating the pemurations. If `TRUE`, the function will fix the random
The default is `FALSE`.}

\item{weighted}{A logical value indicating whether weighted calculations
should be performed. If \code{TRUE}, the function will use limma-based
weights for the calculations.}

\item{propd}{A \code{propd} object.}

\item{what}{A character string. The theta type to set active.}

\item{moderated}{For \code{updateF}, a boolean. Toggles
whether to calculate a moderated F-statistic.}

\item{ivar}{See \code{propr} method.}
}
\value{
A \code{propd} object containing the computed theta values,
 associated count matrix, group labels, and other calculated statistics.
}
\description{
Welcome to the \code{propd} method!

Let \eqn{X} and \eqn{Y} be non-zero positive feature vectors
 measured across \eqn{N} samples belonging to one of two groups,
 sized \eqn{N1} and \eqn{N2}. We use VLR to denote the variance
 of the log of the ratio of the vectors \eqn{X} over \eqn{Y}.
 We define theta as the weighted sum of the within-group VLR
 divided by the weighted total VLR.

The \code{propd} method calculates theta. This fails in
 the setting of zero counts. The \code{propd} method
 will use a Box-Cox transformation to approximate VLR based on
 the parameter \eqn{\alpha}, if provided. We refer the user to
 the vignette for more details.

Note that Group 1 always refers to the first element of the
 \code{group} vector argument supplied to \code{propd}.
}
\details{
The \code{propd} function creates a \code{propd} object, which is
 used for differential analysis of regulatory relationships between features.
 It performs log-ratio transformation, calculates the variance of log-ratio
 (VLR), and computes the theta values using different formulas. The object
 stores the count matrix, group labels, alpha parameter, and other relevant
 information needed for further analysis and visualization.
}
\section{Slots}{

\describe{
\item{\code{counts}}{A data.frame. Stores the original "count matrix" input.}

\item{\code{alpha}}{A double. Stores the alpha value used for transformation.}

\item{\code{group}}{A character vector. Stores the original group labels.}

\item{\code{weighted}}{A logical. Stores whether the theta is weighted.}

\item{\code{weights}}{A matrix. If weighted, stores the limma-based weights.}

\item{\code{active}}{A character. Stores the name of the active theta type.}

\item{\code{Fivar}}{ANY. Stores the reference used to moderate theta.}

\item{\code{dfz}}{A double. Stores the prior df used to moderate theta.}

\item{\code{results}}{A data.frame. Stores the pairwise \code{propd} measurements.}

\item{\code{permutes}}{A data.frame. Stores the shuffled group labels,
used to reproduce permutations of \code{propd}.}

\item{\code{fdr}}{A data.frame. Stores the FDR cutoffs for \code{propd}.}
}}

\section{Methods (by generic)}{

\code{show:} Method to show \code{propd} object.
}

\section{Functions}{

\code{setDisjointed:}
 A wrapper for \code{setActive(propd, what = "theta_d")}.


\code{setEmergent:}
 A wrapper for \code{setActive(propd, what = "theta_e")}.


\code{setActive:}
 Build analyses and figures using a specific theta type. For
 example, set \code{what = "theta_d"} to analyze disjointed
 proportionality and \code{what = "theta_e"} to analyze
 emergent proportionality.


\code{updateF:}
 Use the \code{propd} object to calculate the F-statistic
 from theta as described in the Erb et al. 2017 manuscript
 on differential proportionality. Optionally calculates a
 moderated F-statistic using the limma-voom method. Supports
 weighted and alpha transformed theta values.
}

\examples{
# Sample input count data and group assignments
data <- iris[1:100, 1:4]
group <- iris[1:100, 5]

# Create a propd object for differential analysis
result <- propd(data, group, alpha = 0.5)

}
