% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr.R
\docType{package}
\name{propr}
\alias{propr}
\alias{propr-package}
\title{propr: Ways to measure strength of proportionality.}
\description{
This is the package level documentation for propr and it lives in the file R/propr.R
}
