% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{proprCLR}
\alias{proprCLR}
\title{Calculates the centered log-ratio transformation.}
\usage{
proprCLR(X, check = FALSE)
}
\arguments{
\item{X}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns.}

\item{check}{A logical. If TRUE, function first checks
for negative and NA values.}
}
\value{
A matrix. Returns the centered log-ratio
 transformation of \code{X}.
}
\description{
Provided for backend use.
}
