% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{proprPairs}
\alias{proprPairs}
\title{Recasts proportionality matrix as a table of feature pairs.}
\usage{
proprPairs(prop)
}
\arguments{
\item{prop}{A data.frame or matrix. A proportionality matrix.}
}
\value{
A data.frame. Returns a table of feature pairs.
}
\description{
Provided for backend use.
}
