% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-backend.R
\name{proprPerb}
\alias{proprPerb}
\title{Calculate proportionality metric rho (Erb 2016).}
\usage{
proprPerb(counts, ivar = 0)
}
\arguments{
\item{counts}{A data.frame or matrix. See "count matrix" described in \code{\link{phit}} function details.}

\item{ivar}{A numeric scalar. Specificies feature to use as reference for additive log-ratio transformation.}
}
\value{
Returns proportionality matrix.
}
\description{
Provided for backend use.
}

