% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-backend.R
\name{proprPhit}
\alias{proprPhit}
\title{Calculate proportionality metric phi (Lovell 2015).}
\usage{
proprPhit(counts, symmetrize = TRUE)
}
\arguments{
\item{counts}{A data.frame or matrix. See "count matrix" described in \code{\link{phit}} function details.}

\item{symmetrize}{A logical. If \code{TRUE}, forces symmetry by duplicating the "lower left triangle".}
}
\value{
Returns proportionality matrix.
}
\description{
Provided for backend use.
}

