% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-backend.R
\name{proprSym}
\alias{proprSym}
\title{Symmetrizes a proportionality matrix.}
\usage{
proprSym(prop)
}
\arguments{
\item{prop}{A data.frame or matrix. A proportionality matrix.}
}
\value{
A matrix. Returns a symmetrized proportionality matrix.
}
\description{
Provided for backend use.
}

