% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{proprTri}
\alias{proprTri}
\title{Retrieve the lower left triangle of a proportionality matrix.}
\usage{
proprTri(prop)
}
\arguments{
\item{prop}{A data.frame or matrix. A proportionality matrix.}
}
\value{
A vector. Returns the lower left triangle of a
 proportionality matrix.
}
\description{
Provided for backend use.
}
