% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{proprVLR}
\alias{proprVLR}
\title{Calculates the variance of the log of the ratios.}
\usage{
proprVLR(X, check = FALSE)
}
\arguments{
\item{X}{A data.frame or matrix. A "count matrix" with subjects as rows and features as columns.}

\item{check}{A logical. If TRUE, function first checks for negative and NA values.}
}
\value{
Returns a matrix containing the variance of the log of the ratios.
}
\description{
Provided for backend use.
}

