% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getMatrix.R
\name{results_to_matrix}
\alias{results_to_matrix}
\title{Get Matrix from Results}
\usage{
results_to_matrix(results, what = "theta", features = NULL)
}
\arguments{
\item{results}{A \code{data.frame} of results.}

\item{what}{A character string. The column name of the results data frame to be converted into a matrix.}

\item{features}{A vector of features. Default is NULL.}
}
\value{
A matrix.
}
\description{
This function converts the results data frame into a matrix.
}
