% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-graflex.R
\name{runGraflex}
\alias{runGraflex}
\title{Calculate odds ratio and FDR}
\usage{
runGraflex(A, K, p = 100, ncores = 1)
}
\arguments{
\item{A}{An adjacency matrix.}

\item{K}{A knowledge database where each row is a graph node
and each column is a concept.}

\item{p}{An integer. The number of permutation.}
}
\description{
This function calls \code{\link{graflex}} for each
concept (i.e., column) in the database \code{K}.
}
\details{
For each concept, this function calculates the odds ratio
and determines the false discovery rate (FDR) by counting 
the number of the actual OR was greater or less than a 
permuted OR.
}
