% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c-propd-experimental.R
\name{runNormalization}
\alias{runNormalization}
\title{Get Per-Feature Theta}
\usage{
runNormalization(object, norm.factors)
}
\arguments{
\item{object}{A \code{\link{propd}} object.}

\item{norm.factors}{A numeric vector. The effective library size
normalization factors (e.g., from edgeR or DESeq2).}
}
\value{
A numeric vector. A theta for each feature.
}
\description{
This function calculates the differential proportionality
 between each feature and a set of normalization factors. When the
 normalization factors correctly remove the compositional bias, the
 resultant thetas indicate differential expression (DE). However, unlike
 other DE tests, the p-value for differential proportionality is
 not linked to the normalization factors. Here, normalization factors
 only affect the interpretation, not the statistics.
}
