% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1a-propr-backend.R
\name{simple_zero_replacement}
\alias{simple_zero_replacement}
\title{Simple Zero Replacement in a Count Matrix}
\usage{
simple_zero_replacement(ct)
}
\arguments{
\item{ct}{A data matrix for which the log-ratio transformation will be
performed. It is assumed that the matrix contains numerical values only.}
}
\value{
A matrix with zero values replaced by the next smallest non-zero value.
}
\description{
This function replaces zeros with the next smallest non-zero value in the
 input count matrix. If the matrix contains no zeros, it produces an
 informational message indicating that no replacements were made.
}
\examples{
# Sample input count data with zeros
data <- matrix(c(0, 2, 3, 4, 5, 0), nrow = 2, byrow = TRUE)

}
