% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify Indexed Matrix}
\usage{
simplify(object)
}
\arguments{
\item{object}{An object of class \code{propr}.}
}
\value{
Returns a subsetted and indexed \code{propr} object.
}
\description{
This convenience function takes an indexed \code{\link{propr}} object and
 subsets the object based on that index. In addition, it populates
 the \code{@pairs} slot of the new object with an updated version
 of the original index.
}

