% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-functions.R
\name{sma}
\alias{sma}
\title{Standardised Major Axis fits of pairs of columns}
\usage{
sma(X)
}
\arguments{
\item{X}{A matrix or dataframe}
}
\value{
A list of three elements \code{b}, \code{p} and \code{r2}.
}
\description{
\code{sma} returns a list whose elements are matrices whose elements (i,j)
relate to the Standardised Major Axis fits of columns i and j of \code{X}
}
\details{
\strong{Note:} \code{X} should be the result of a centred logratio transformation
}
\examples{
N <- 100 # Number of observations
# Make a data frame with columns a and b roughly proportional
# and columns c and d unrelated
a <- seq(from=5, to=15, len=N)
b <- a * rnorm(N, 1, 0.1)
c <- rnorm(N,10)
d <- rnorm(N,10)
X <- data.frame(a, b, c, d)
pairs(X)
pairs(clr(X)) # Note the spurious correlation between variables c and d
sma(clr(X))
}
