% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aware.R
\name{smear}
\alias{smear}
\title{Make Smear Plot}
\usage{
smear(rho, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.
Note that some exceptions apply.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}
}
\value{
Returns a \code{ggplot} object.
}
\description{
Plots *lr-transformed abundances for all indexed pairs.
}
\details{
Note that only \code{\link{smear}} and \code{\link{dendrogram}}
 produce "index-aware" plots. These functions also accommodate
 results from either \code{\link{phit}} or \code{\link{perb}}.
}

