% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{top.counts}
\alias{top.counts}
\title{Example propr Object}
\format{An object of class \code{matrix} with 20 rows and 409 columns.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(top.counts)
}
\description{
Includes the non-transformed count abundances from all
 cane toad transcripts with at least 10 counts in at least
 10 samples, subsetted to include only those indexed
 by \code{rho > .995}. Used for vignette.
}
\keyword{datasets}
