% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{top}
\alias{top}
\title{Example propr Object}
\format{An object of class \code{propr} of length 1.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(top)
}
\description{
Includes cane toad transcripts with at least 10 counts
 across at least 10 samples. Used for vignette.
}
\keyword{datasets}

