% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{updateCutoffs_old.propr}
\alias{updateCutoffs_old.propr}
\title{Old version of updateCutoffs.propr}
\usage{
updateCutoffs_old.propr(object, cutoff = seq(0.05, 0.95, 0.3))
}
\description{
Do not use this function. For testing purposes only.
}
