% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-functions.R
\name{vlr}
\alias{vlr}
\title{Variance of logratios}
\usage{
vlr(X, check = FALSE)
}
\arguments{
\item{X}{A matrix or dataframe of positive numeric values}

\item{check}{A logical scalar}
}
\value{
The symmetric matrix
\eqn{\mathrm{Var}{\log(X_i/X_j)}}{Var(log(X_i/X_j))} where \eqn{X_i} and \eqn{X_j}
denote \emph{columns} \eqn{i} and \eqn{j} of \eqn{X}.
}
\description{
\code{vlr} returns a matrix where element (i,j) is
the variance (over rows) of the log of the ratios of column i and j.
}
\details{
If \code{check} is \code{TRUE} then this function will stop if
there are any negative or \code{NA} values in \code{X}.
}
\examples{
N <- 10 # Number of observations
# Make a data frame with columns a and b roughly proportional
# and columns c and d roughly proportional
X <- data.frame(a=(1:N), b=(1:N) * rnorm(N, 10, 0.1),
                c=(N:1), d=(N:1) * rnorm(N, 10, 1.0))
round(vlr(X),2)
}
