% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2prob.R
\name{abstract}
\alias{abstract}
\title{Abstract Two propr Objects}
\usage{
abstract(x, y, select)
}
\arguments{
\item{x, y}{A \code{propr} object.}

\item{select}{Subsets via \code{object@counts[, select]}.
Optional. Use this argument to subset the proportionality
matrix without altering the final value of \code{rho}.}
}
\value{
An abstracted \code{propr} object.
}
\description{
This function abstracts a new \code{propr} object from
 two existing \code{propr} objects. The two \code{propr}
 objects should not have any overlapping samples. Typically,
 the two objects represent different experimental groups.
 The resultant abstracted object inherits all plot functions
 available for the original \code{propr} objects.
}
\details{
The abstracted \code{propr} object has the following properties:
 The \code{@counts} and \code{@logratio} slots contain a
 join of the original slots via \code{rbind}. Meanwhile,
 the \code{@matrix} slot contains a difference matrix defined as
 \code{1 - abs(x@matrix - y@matrix)}. This difference matrix
 may help summarize \code{\link{prop2prob}} results.
}
\examples{
library(propr)
data(mail)
mail1 <- mail[1:2, ]
mail2 <- mail[3:4, ]
rho1 <- perb(mail1)
rho2 <- perb(mail2)
abstract(rho1, rho2)
}
\seealso{
\code{\link{propr}}, \code{\link{prop2prob}}
}

