% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{adjacent}
\alias{adjacent}
\title{Make Adjacency Object}
\usage{
adjacent(object)
}
\arguments{
\item{object}{An object of class \code{propr}.}
}
\value{
Returns a \code{propr} object with the adjacency
 matrix saved to the \code{@matrix} slot.
}
\description{
This function uses pairs indexed in the \code{@pairs}
 slot to build a symmetric adjacency matrix.
}

