% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2prob.R
\name{adjacent}
\alias{adjacent}
\title{Make Adjacency Object}
\usage{
adjacent(rho)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.
Note that some exceptions apply.}
}
\value{
Returns a \code{propr} object with the adjacency
 matrix saved to the \code{@matrix} slot.
}
\description{
This function uses pairs indexed in the \code{@pairs}
 slot to build a symmetric adjacency matrix.
}

