% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex2propr.R
\name{aldex.cor}
\alias{aldex.cor}
\title{Correlate CLR with a Continuous Measurement}
\usage{
aldex.cor(clr, conditions, ...)
}
\arguments{
\item{clr}{An \code{aldex.clr} object.}

\item{conditions}{A numeric vector of a continuous variable.}

\item{...}{Arguments passed to \code{cor.test}.}
}
\value{
Returns a data.frame of the average
 correlation statistic (e.g., \code{r}) and
 average p-value (\code{p}) for each feature
 across all Monte Carlo instances. Average
 FDR provided by \code{BH} column.
}
\description{
This function uses the Monte Carlo instances from an
 \code{aldex.clr} object to correlate each log-ratio
 transformed feature vector with a continuous
 numeric variable. See \code{\link{lr2cor}}.
}
