% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{caneToad.counts}
\alias{caneToad.counts}
\title{Example Cane Toad Count Data}
\format{An object of class \code{matrix} with 20 rows and 57580 columns.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(caneToad.counts)
}
\description{
Example Cane Toad Count Data
}
\keyword{datasets}

