% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{caneToad.groups}
\alias{caneToad.groups}
\title{Example Cane Toad Group Data}
\format{An object of class \code{character} of length 20.}
\source{
<DOI:10.1111/mec.13184>
}
\usage{
data(caneToad.groups)
}
\description{
Group labels for twenty toads sampled from one
 of two regions in Australia. Data set reduced to
 exclude any transcripts without at least 10 counts
 in at least 10 samples.
}
\keyword{datasets}
