% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dendrogram}
\alias{dendrogram}
\title{Make Dendrogram Plot}
\usage{
dendrogram(rho, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}
}
\value{
Returns a dendrogram object made from \code{hclust}.
}
\description{
Plots a dendrogram and heatmap for all indexed pairs.
}

