% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-aware.R
\name{dendrogram}
\alias{dendrogram}
\title{Make Dendrogram Plot}
\usage{
dendrogram(rho, prompt = TRUE, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.
However, \code{\link{smear}} and \code{\link{dendrogram}} will also
accommodate a \code{propr} object created from \code{\link{phit}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the courtesy prompt when working with big data.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}
}
\value{
Returns a dendrogram object made from \code{hclust}.
}
\description{
Plots a dendrogram and heatmap for all indexed pairs.
}
\details{
Note that only \code{\link{smear}} and \code{\link{dendrogram}}
 produce "index-aware" plots. These functions also accommodate
 results from either \code{\link{phit}} or \code{\link{perb}}.
}

