% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-get.R
\name{getAdj}
\alias{getAdj}
\title{Get Object as Adjacency Matrix}
\usage{
getAdj(object, cutoff = 1)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}

\item{cutoff}{This argument indicates the value at which to
cutoff the results. For "rho" and "cor", the function
returns pairs with a value greater than the cutoff.
For "theta", "phi", and "phs", the function returns pairs
with a value less than the cutoff. Leave the argument as
\code{NA} to return all results.}
}
\value{
An adjacency matrix.
}
\description{
This function uses \code{getMatrix} to
 return a \code{propr} or \code{propd}
 object as an adjacency matrix.
 The diagonal is set to 0.
}
\details{
This function is faster than
 \code{getAdjacency}, but has fewer
 arguments to tailor results.
}
