% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getAdjacency.R
\name{getAdjacencyFDR}
\alias{getAdjacencyFDR}
\alias{getAdjacencyFDR.propd}
\alias{getAdjacencyFDR.propr}
\title{Get Adjacency Matrix as indicated by permutation tests}
\usage{
getAdjacencyFDR(object, fdr = 0.05, window_size = 1, tails = NULL)

getAdjacencyFDR.propd(object, fdr = 0.05, window_size = 1)

getAdjacencyFDR.propr(object, fdr = 0.05, window_size = 1, tails = NULL)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{fdr}{A float value for the false discovery rate. Default is 0.05.}

\item{window_size}{An integer. Default is 1. When it is greater than 1, the FDR
values would be smoothed out by a moving average of the given window size.}

\item{tails}{1 for one-sided on the right, and 2 for two-sided. When NULL, use default
test for the given metric. This is only relevant for \code{propr} objects.}
}
\value{
An adjacency matrix.
}
\description{
This function gets the significant pairs, according to the permutation tests. 
Then it fills the adjacency matrix with 1 if pair is significant, otherwise 0.
}
\section{Methods}{

\code{getAdjacencyFDR.propd:}
This function creates an adjacency matrix with the significant
pairs from a \code{propd} object.


\code{getAdjacencyFDR.propr:}
This function creates an adjacency matrix with the significant
pairs from a \code{propd} object.
}

