% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getAdjacency.R
\name{getAdjacencyFDR}
\alias{getAdjacencyFDR}
\title{Get Adjacency Matrix as indicated by permutation tests.}
\usage{
getAdjacencyFDR(object, fdr = 0.05, window_size = 1)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{fdr}{A float value for the false discovery rate. Default is 0.05.}

\item{window_size}{An integer. Default is 1. When it is greater than 1, the FDR
values would be smoothed out by a moving average of the given window size.}
}
\value{
An adjacency matrix.
}
\description{
This function gets the significant pairs according to the permutation tests. Then it fills 
the adjacency matrix with 1 if pair is significant, otherwise 0. The significance is determined
by the cutoff value for which the false discovery rate (FDR) is less or equal than the given 
value 'fdr'. The significant pairs are those that have a value greater/less or equal than the 
cutoff, depending on the metric.
}
