% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getAdj.R
\name{getAdjFDR}
\alias{getAdjFDR}
\title{Get Adjacency Matrix as indicated by permutation tests}
\usage{
getAdjFDR(object, fdr = 0.05, positive = TRUE, negative = FALSE)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{fdr}{A float value for the false discovery rate. Default is 0.05.}

\item{positive}{A boolean. In the case of \code{propr} object,
toggles where to retain the significant pairs with positive values (>=0).}

\item{negative}{A boolean. In the case of \code{propr} object,
toggles where to retain the significant pairs with negative values (<0).}
}
\value{
An adjacency matrix.
}
\description{
This function gets the significant pairs, according to the permutation tests. 
Then it fills the adjacency matrix with 1 if pair is significant, otherwise 0.
The diagonal is set to 0.
}
