% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getCutoff.R
\name{getCutoffFDR}
\alias{getCutoffFDR}
\title{Get a meaningful cutoff based on the FDR values from permutation tests.}
\usage{
getCutoffFDR(object, fdr = 0.05, positive = TRUE)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{fdr}{A float value for the false discovery rate.
Default is 0.05.}

\item{positive}{A boolean. Toggles whether to return the cutoff with
positive value. Otherwise, return cutoff with negative value.}
}
\value{
A cutoff value.
}
\description{
Get a meaningful cutoff based on the FDR values from permutation tests.
}
