% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-get.R
\name{getMatrix}
\alias{getMatrix}
\title{Get Object as Symmetric Matrix}
\usage{
getMatrix(object)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}
}
\value{
A symmetric matrix.
}
\description{
This function returns a symmetric matrix
 of \code{propr} or \code{propd} values.
 The diagonal is set to 0.
}
