% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getResults.R
\name{getResults}
\alias{getResults}
\title{Get Results from Object}
\usage{
getResults(object)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}
}
\value{
A \code{data.frame} of results.
}
\description{
This function provides a unified wrapper to retrieve results
 from a \code{propr} or \code{propd} object.
}
