% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getResults.R
\name{getSignificantResultsFDR}
\alias{getSignificantResultsFDR}
\alias{getSignificantResultsFDR.propr}
\alias{getSignificantResultsFDR.propd}
\title{Get Significant Results from Object based on the permutation tests.}
\usage{
getSignificantResultsFDR(object, fdr = 0.05, window_size = 1)

getSignificantResultsFDR.propr(object, fdr = 0.05, window_size = 1)

getSignificantResultsFDR.propd(object, fdr = 0.05, window_size = 1)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{fdr}{A float value for the false discovery rate. Default is 0.05.}

\item{window_size}{An integer. Default is 1. When it is greater than 1, the FDR
values would be smoothed out by a moving average of the given window size.}
}
\value{
A \code{data.frame} of results.
}
\description{
This function retrieves results from a \code{propr} or \code{propd} object keeping only the
statistically significant pairs. The significance is determined by the cutoff value for which
the false discovery rate (FDR) is less or equal than the given value 'fdr'. The significant 
pairs are those that have a value greater/less or equal than the cutoff, depending on the metric.
}
\section{Methods}{

\code{getSignificantResultsFDR.propr:}
This function retrieves results from a \code{propr} object keeping 
only the statistically significant pairs.


\code{getSignificantResultsFDR.propd:}
This function retrieves results from a \code{propd} object keeping 
only the statistically significant pairs.
}

