% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getResults.R
\name{getSignificantResultsFDR}
\alias{getSignificantResultsFDR}
\title{Get Significant Results from Object based on the permutation tests.}
\usage{
getSignificantResultsFDR(object, fdr = 0.05, positive = TRUE, negative = FALSE)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}

\item{fdr}{A numeric. The false discovery rate to use.}

\item{positive}{A boolean. In the case of \code{propr} object,
toggles whether to return the significant positive values (>=0).}

\item{negative}{A boolean. In the case of \code{propr} object,
toggles whether to return the significant negative values (<0).}
}
\value{
A \code{data.frame} of results.
}
\description{
This function provides a unified wrapper to retrieve results
 from a \code{propr} or \code{propd} object keeping only the
 statistically significant pairs.
}
