% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-getResults.R
\name{getSignificantResultsFstat}
\alias{getSignificantResultsFstat}
\title{Get Significant Results based on the F-stats.}
\usage{
getSignificantResultsFstat(object, pval = 0.05, fdr_adjusted = TRUE)
}
\arguments{
\item{object}{A \code{propd} or \code{propr} object.}

\item{pval}{A float value for the p-value. Default is 0.05.}

\item{fdr_adjusted}{A boolean. If TRUE, use the the FDR- adjusted p-values. 
Otherwise, get significant pairs based on the theoretical F-statistic cutoff.}
}
\value{
A \code{data.frame} of results.
}
\description{
This function provides a unified wrapper to retrieve results
 from a \code{propd} object keeping only the statistically 
 significant pairs. Note that it can only be applied to theta_d,
 as updateF only works for theta_d.
}
