% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-aldex2propr.R
\name{lr2cor}
\alias{lr2cor}
\title{Correlate CLR Data with a Continuous Measurement}
\usage{
lr2cor(lr, conditions, ...)
}
\arguments{
\item{lr}{A data.frame. A log-ratio transformed counts matrix.}

\item{conditions}{A numeric vector of a continuous variable.}

\item{...}{Arguments passed to \code{cor.test}.}
}
\value{
Returns a data.frame of the estimated
 correlation statistic (e.g., \code{r}) and p-value
 (\code{p}) for each log-ratio transformed feature,
 with FDR appended as the \code{BH} column.
}
\description{
This function correlates each log-ratio transformed
 feature vector with a continuous numeric variable.
}
