% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex.cor.R
\name{lr2cor}
\alias{lr2cor}
\title{Correlate CLR with a Continuous Measurement}
\usage{
lr2cor(lr, conditions, ...)
}
\arguments{
\item{lr}{A data.frame. A log-ratio transformed counts matrix.}

\item{conditions}{A numeric vector of a continuous variable.}

\item{...}{Arguments passed to \code{cor.test}.}
}
\value{
Returns a data.frame of the correlation
 statistic (e.g., \code{r}) and p-value (\code{p})
 for each log-ratio transformed feature.
 FDR provided by \code{BH} column.
}
\description{
This function correlates each log-ratio transformed
 feature vector with a continuous numeric variable.
}

