% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mds}
\alias{mds}
\title{Make MDS Plot}
\usage{
mds(rho, group, prompt = TRUE, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the column names in \code{@counts} and
\code{@logratio}. Parameter required for \code{\link{bucket}}
and optional for \code{\link{mds}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the prompt when plotting a large number of features.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package. This will also
have the function return the internal \code{ggplot} object(s),
useful for customization of the figure.}
}
\description{
Plots the first two principal components as calculated
 using the log-ratio transformed feature vectors. This
 provides a statistically valid alternative to
 conventional principal components analysis (PCA) and
 multi-dimensional scaling (MDS) plotting.
}
\details{
For more information, see DOI:10.1139/cjm-2015-0821.
}

