% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propdViz.R
\name{pals}
\alias{pals}
\title{Calculate PALs for Pairs}
\usage{
pals(object, k)
}
\arguments{
\item{object}{A \code{propd} object.}

\item{k}{An integer. The maximum number of PALs to index
when calculating \code{\link{pals}} in the network.}
}
\value{
A named vector of PALs, ordered by decreasing
 connectivity of the input nodes. The names of the result
 refer to the input nodes themselves.
}
\description{
This function finds the Popular Adjacent Ligand or Self (PALs)
 for each feature in the \code{@theta} slot of a \code{propd}
 object. Specifically, we define PALs as the adjacent node with
 the highest amount of connectivity. If node itself has more
 connectivity than any of its neighbors, it is its own PAL.
}
