% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated2.R
\name{permuteTheta_naive}
\alias{permuteTheta_naive}
\title{Permute Theta}
\usage{
permuteTheta_naive(counts, group, p = 64)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns.}

\item{group}{A character vector. Group or sub-group memberships,
ordered according to the row names in \code{counts}.}

\item{p}{An integer. The number of permutation cycles.}
}
\description{
Permute differential proportionality measure, theta.
}
\details{
This function randomizes all feature vectors \code{p} times.
}
