% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCheck}
\alias{plotCheck}
\title{Plot Check}
\usage{
plotCheck(x)
}
\arguments{
\item{x}{An object of class \code{propr}.}
}
\description{
Performs data checks before plotting, triggering messages
 or errors when appropriate. For back-end use only.
}

