% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCheck}
\alias{plotCheck}
\title{Plot Check}
\usage{
plotCheck(rho, prompt)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the prompt when plotting a large number of features.}
}
\description{
Performs data checks before plotting, triggering messages
 or errors when appropriate. For back-end use only.
}

