% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{plotCheck}
\alias{plotCheck}
\title{Plot Check}
\usage{
plotCheck(rho, prompt, plotly, indexNaive)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{perb}.
However, \code{smear}, \code{dendrogram}, and \code{cytescape} will
also accommodate results from \code{phit} and \code{phis}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the courtesy prompt when working with big data.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}

\item{indexNaive}{Toggles whether to perform checks for an
"index-naive" plot function.}
}
\value{
Returns a \code{propr} object with guaranteed
 column names and row names.
}
\description{
Performs data checks before plotting, triggering messages
 or errors when appropriate. For back-end use only.
}

