% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-propd-genewise.R
\name{propdGenewise}
\alias{propdGenewise}
\title{Convert pairwise propd results into genewise}
\usage{
propdGenewise(
  propd,
  metric = c("connectivity", "wconnectivity"),
  pairwise_fdr = 0.05
)
}
\arguments{
\item{propd}{A \code{\link{propd}} object, with FDR values from updateF. Note:
for the moment, only theta results with F-stats are supported. Later we will look 
on the option to get FDR values based on permutations.}

\item{metric}{A character string. The metric to use for genewise results. Options 
are "connectivity" and "wconnectivity".}

\item{pairwise_fdr}{}
}
\value{
A data frame with genewise results.
}
\description{
This function converts pairwise propd results into genewise results. The resulting
genewise results are direct indirectors of genes being differentially expressed.
}
\details{
The "connectivity" metric refers to the number of significant pairwise 
relationships a gene has, while the "wconnectivity" metric weights these 
relationships by their strength (i.e., theta values). The resulting genewise results 
can be used to identify genes that are central to the observed proportionality changes 
across groups (these genes tend to be differentially expressed).
}
