% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{propr-class}
\alias{propr-class}
\title{An S4 class to hold results from proportionality analysis.}
\description{
An S4 class to hold results from proportionality analysis.
}
\section{Slots}{

\describe{
\item{\code{counts}}{A matrix. Stores the original "count matrix" input.}

\item{\code{logratio}}{A matrix. Stores the log-ratio transformed "count matrix".}

\item{\code{matrix}}{A matrix. Stores the proportionality matrix calculated by \code{phit} or \code{perb}.}

\item{\code{pairs}}{A vector. Indexes the proportionality metrics of interest.}
}}
\examples{
randomNum <- sample(1:1000, size = 25 * 10, replace = TRUE)
counts <- matrix(randomNum, nrow = 25, ncol = 10)
prop <- perb(counts, ivar = 0)
prop[">", .75]
subset(prop, 1:5)
}
\seealso{
\code{\link{propr}}, \code{\link{phit}}, \code{\link{perb}}
}

