% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propr-classes.R, R/propr-methods.R
\docType{class}
\name{propr-class}
\alias{$,propr-method}
\alias{[,propr-method}
\alias{propr-class}
\alias{show,propr-method}
\alias{subset,propr-method}
\title{An S4 class to hold results from proportionality analysis.}
\usage{
\S4method{show}{propr}(object)

\S4method{subset}{propr}(x, subset)

\S4method{[}{propr}(x, i, j, drop)

\S4method{$}{propr}(x, name)
}
\arguments{
\item{object, x}{An object of class \code{propr}.}

\item{subset}{Subsets via \code{object@counts[subset, ]}.}

\item{i, j, drop}{Subsets via \code{object@pairs[i, j, drop]}.}

\item{name}{Subsets via \code{object@pairs[, name]}.}
}
\description{
An S4 class to hold results from proportionality analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{propr} object.

\item \code{subset}: Method to subset \code{propr} object.

\item \code{[}: Method to subset \code{propr} object.

\item \code{$}: Method to subset \code{propr} object.
}}
\section{Slots}{

\describe{
\item{\code{counts}}{A data.frame. Stores the original "counts matrix" input.}

\item{\code{logratio}}{A data.frame. Stores the log-ratio transformed "counts matrix".}

\item{\code{matrix}}{A matrix. Stores the proportionality matrix calculated by \code{phit} or \code{perb}.}

\item{\code{pairs}}{A data.frame. Projects the proportionality matrix pairwise.}
}}
\examples{
randomNum <- sample(1:1000, size = 2000 * 22, replace = TRUE)
counts <- matrix(randomNum, nrow = 2000, ncol = 22)
prop <- perb(counts, ivar = 0, iter = 0)
prop[1:5, ]
prop$prop
prop[1:5, "prop"]
subset(prop, 1:5)
}
\seealso{
\code{\link{phit}}, \code{\link{perb}}
}

