% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/methods.R, R/plot.R
\docType{methods}
\name{propr}
\alias{[,propr,ANY,ANY-method}
\alias{[,propr-method}
\alias{plot,propr,missing-method}
\alias{propr}
\alias{show,propr-method}
\alias{subset,propr-method}
\title{The propr Package}
\usage{
\S4method{show}{propr}(object)

\S4method{subset}{propr}(x, subset, select)

\S4method{[}{propr,ANY,ANY}(x, i = "all", j, tiny = FALSE)

\S4method{plot}{propr,missing}(x, y, plotly = FALSE)
}
\arguments{
\item{object, x}{An object of class \code{propr}.}

\item{subset}{Subsets via \code{object@counts[subset, ]}.
Use this argument to rearrange subject order.}

\item{select}{Subsets via \code{object@counts[, select]}.
Use this argument to rearrange feature order.}

\item{i}{Operation used for the subset indexing. Select from
"==", "=", ">", ">=", "<", "<=", "!=", or "all".}

\item{j}{Reference used for the subset indexing. Provide a numeric
value to which to compare the proportionality measures in the
\code{@matrix} slot.}

\item{tiny}{A logical scalar. Toggles whether to pass the indexed
result through \code{\link{simplify}}.}

\item{y}{Missing. Ignore. Leftover from the generic method definition.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}
}
\description{
Welcome to the \code{propr} package!

To learn more about how to calculate proportionality, see
 see \code{\link{phit}} and \code{\link{perb}}.

To learn more about the resultant \code{propr} class object, see
 see the class definition \code{\link{propr-class}}.

To learn more about \code{propr} plots, see \code{\link{smear}},
 \code{\link{dendrogram}}, \code{\link{bucket}}, \code{\link{prism}},
 \code{\link{bokeh}}, \code{\link{mds}}, and \code{\link{snapshot}}.

To learn more about differential proportionality, see
 \code{\link{prop2prob}} and \code{\link{abstract}}.

To learn more about compositional data analysis, and its relevance
 to biological count data, see the bundled vignette.
}
\section{Methods (by generic)}{

\code{show:} Method to show \code{propr} object.


\code{subset:} Method to subset \code{propr} object.


\code{[:} Method to subset \code{propr} object.


\code{plot:} Method to plot \code{propr} object.
}

