% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{proprPerb}
\alias{proprPerb}
\title{Calculate proportionality metric rho (Erb 2016).}
\usage{
proprPerb(counts, ivar = 0)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns.}

\item{ivar}{A numeric scalar. Specifies reference feature(s)
for additive log-ratio transformation. The argument will also
accept feature name(s) instead of the index position(s).
Set to "iqlr" to use inter-quartile log-ratio transformation.
Ignore to use centered log-ratio transformation.}
}
\value{
Returns proportionality matrix.
}
\description{
Provided for backend use.
}
