% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-deprecated.R
\name{proprPhit}
\alias{proprPhit}
\title{Calculate proportionality metric phi (Lovell 2015).}
\usage{
proprPhit(counts, symmetrize = TRUE)
}
\arguments{
\item{counts}{A data.frame or matrix. A "count matrix" with
subjects as rows and features as columns. Note that this matrix
does not necessarily have to contain counts.}

\item{symmetrize}{A logical. If \code{TRUE}, forces symmetry
by reflecting the "lower left triangle".}
}
\value{
Returns proportionality matrix.
}
\description{
Provided for backend use.
}
