% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-propdCall.R
\name{qtheta}
\alias{qtheta}
\title{Calculate a theta Cutoff}
\usage{
qtheta(propd, moderated = FALSE, pval = 0.05)
}
\arguments{
\item{propd}{A \code{propr} or \code{propd} object.}

\item{moderated}{For \code{updateF}, a boolean. Toggles
whether to calculate a moderated F-statistic.}

\item{pval}{A p-value at which to calculate a theta cutoff.}
}
\value{
A cutoff of theta from [0, 1].
}
\description{
This function uses the F distribution to calculate a cutoff of
 theta for a p-value given by the \code{pval} argument.
}
