% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{ratios}
\alias{ratios}
\title{Recast Matrix as Feature Ratios}
\usage{
ratios(matrix)
}
\arguments{
\item{matrix}{A matrix. The data to recast.}
}
\value{
A matrix.
}
\description{
The \code{ratios} function recasts a matrix with N feature columns
 as a new matrix with N * (N - 1) / 2 feature ratio columns.
}
