% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d-propd-graflex.R
\name{runGraflex}
\alias{runGraflex}
\title{Permute FDR for Multiple Concepts}
\usage{
runGraflex(A, K, p = 500)
}
\arguments{
\item{A}{An adjacency matrix.}

\item{K}{A knowledge database where each row is a graph node
and each column is a concept.}

\item{p}{An integer. The number of overlaps to permute.}
}
\description{
This function calls \code{\link{permuteOR}} for each
 concept (i.e., column) in the database \code{K}.
}
\details{
For each concept, this function calculates the
 false discovery rate (FDR) by counting the number of
 times the actual OR was greater than
 (or less than) a permuted OR.
}
