% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{slate}
\alias{slate}
\title{Build \code{propr} Table}
\usage{
slate(rho, k, prompt = TRUE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{k}{A numeric scalar. The number of groups into which to
cluster the subjects. Clusters calculated  using \code{hclust}.
Optional parameter for \code{\link{bucket}}, \code{\link{prism}},
and \code{\link{bokeh}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the prompt when plotting a large number of features.}
}
\value{
Returns a \code{data.frame} of pairwise relationships.
 If the argument \code{k} is provided, returns a list of
 the \code{data.frame} of pairwise relationships and the
 cluster membership.
}
\description{
This function builds a table of VLR, VLS, and rho
 for each feature pair in a \code{propr} object. If the
 argument \code{k} is provided, the table will also
 include co-cluster membership (as matching the
 \code{\link{bucket}} or \code{\link{prism}} plots).
}

