% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{smear}
\alias{smear}
\title{Make Smear Plot}
\usage{
smear(rho, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package. This will also
have the function return the internal \code{ggplot} object(s),
useful for customization of the figure.}
}
\description{
Plots *lr-transformed abundances for all indexed pairs.
}

