% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{snapshot}
\alias{snapshot}
\title{Make Snapshot Plot}
\usage{
snapshot(rho, prompt = TRUE, plotly = FALSE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the prompt when plotting a large number of features.}

\item{plotly}{A logical scalar. Set to \code{TRUE} to produce
a dynamic plot using the \code{plotly} package.}
}
\value{
Returns a dendrogram object made from \code{hclust}.
}
\description{
Plots the log-ratio transformed feature abundance as
 a heatmap, along with the respective dendrograms.
 Heatmap intensity is not scaled.
}

