% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{snapshot}
\alias{snapshot}
\title{Make Snapshot Plot}
\usage{
snapshot(rho, prompt = TRUE)
}
\arguments{
\item{rho}{A \code{propr} object created from \code{\link{perb}}.}

\item{prompt}{A logical scalar. Set to \code{FALSE} to disable
the prompt when plotting a large number of features.}
}
\description{
Plots the log-ratio transformed feature abundance as
 a heatmap, with the axes ordered by dendrogram. Heatmap
 intensity is scaled across the feature vector to make it
 easier to assess visually the difference in the
 log-ratio transformed abundance between samples.
}

