% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{updateCutoffs}
\alias{updateCutoffs}
\title{Update FDR by Permutation}
\usage{
updateCutoffs(object, cutoff = seq(0.05, 0.95, 0.3), ncores = 1)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}

\item{cutoff}{For \code{updateCutoffs}, a numeric vector.
this argument provides the FDR cutoffs to test.
For graph functions, a numeric scalar. This argument
indicates the maximum theta to include in the figure.
For graph functions, a large integer will instead
retrieve the top N pairs as ranked by theta.}

\item{ncores}{An integer. The number of parallel cores to use.}
}
\value{
A \code{propr} or \code{propd} object.
}
\description{
This function updates FDR for a set of cutoffs.
}
\details{
This function wraps \code{updateCutoffs.propr} and
 \code{updateCutoffs.propd}.
}
