% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-shared-updatePermutes.R
\name{updatePermutes}
\alias{updatePermutes}
\title{Create permuted data}
\usage{
updatePermutes(
  object,
  p = 100,
  permutation_option = c("feature-wise", "sample-wise")
)
}
\arguments{
\item{object}{A \code{propr} or \code{propd} object.}

\item{p}{The number of permutations to perform. Default is 100.}

\item{permutation_option}{A character string indicating if permute the data
sample-wise or feature-wise. Default is "feature-wise". Note that this flag
is only relevant for \code{propr} objects.}
}
\value{
A \code{propr} or \code{propd} object with the permutes slot updated.
}
\description{
This function creates p permuted data matrices
}
\details{
This function wraps \code{updatePermutes.propr} and
 \code{updatePermutes.propd}.
}
